Fix right-to-left text selection
Two PDF readers I've tried (Firefox and koreader) both select text
incorrectly if its rendered starting from left columns as Mokuro2Pdf
previously did, starting from the left column. When copying these
sentences to a dictionary, the result is a sentence with text transposed
in chunks. This behavior is corrected if the PDF is rendered right to
left.